requirements.txt
# 核心依赖
pandas>=1.5.0
numpy>=1.21.0
asyncio
pathlib

# 网络请求
httpx>=0.24.0
openai>=1.0.0

# 数据处理
akshare>=1.9.0
matplotlib>=3.5.0
loguru>=0.7.0

# 配置文件处理
pyyaml>=6.0

# 其他工具
tiktoken>=0.5.0
